refactor: use direct imports where possible
