+++ src/DataTable/ToolbarMenuItem.svelte
@@ -1,7 +1,7 @@
 <script>
   /** @extends {"../OverflowMenu/OverflowMenuItem.svelte"} OverflowMenuItemProps */
 
-  import { OverflowMenuItem } from "../OverflowMenu";
+  import OverflowMenuItem from "../OverflowMenu/OverflowMenuItem.svelte";
 </script>
 
 <OverflowMenuItem {...$$restProps} on:click on:keydown>
