+++ src/FileUploader/FileUploaderSkeleton.svelte
@@ -1,6 +1,6 @@
 <script>
-  import { ButtonSkeleton } from "../Button";
-  import { SkeletonText } from "../SkeletonText";
+  import ButtonSkeleton from "../Button/ButtonSkeleton.svelte";
+  import SkeletonText from "../SkeletonText/SkeletonText.svelte";
 </script>
 
 <!-- svelte-ignore a11y-mouse-events-have-key-events -->
